refactor: NetworkConnectionUtil.kt util 패키지로 이동
#186
