+++ app/src/main/java/com/boosters/promise/data/member/source/remote/MemberRemoteDataSource.kt
@@ -1,7 +1,7 @@
 package com.boosters.promise.data.member.source.remote
 
 import com.boosters.promise.data.member.di.MemberModule
-import com.boosters.promise.data.network.NetworkConnectionUtil
+import com.boosters.promise.util.NetworkConnectionUtil
 import com.google.firebase.firestore.CollectionReference
 import com.google.firebase.firestore.ktx.snapshots
 import kotlinx.coroutines.flow.Flow
